Reorganizing the interface for frost
